devide core logic
