+++ lib/kuromoji/core.rb
@@ -1,0 +1,61 @@
+#!/usr/bin/env ruby
+# coding: utf-8
+require 'rjb'
+class Rjb::Rjb_JavaProxy
+  def to_kuromoji_hash
+    result = {}
+    result['parts_of_speech'] = part_of_speech
+    result['reading'] = getReading
+    result['base_form'] = base_form
+    result['surface_form'] = surface_form
+    result['position'] = position
+    result['is_known'] = isKnown
+    result['is_user'] = isUser
+    result
+  end
+end
+
+module Kuromoji
+  class Core
+
+    def initialize(user_dictionary = nil)
+      jar = File.expand_path('../../../vendor/kuromoji-0.7.7/lib/kuromoji-0.7.7.jar', __FILE__)
+      Rjb.load(jar)
+      if user_dictionary.nil?
+        @tokenizer = Rjb.import('org.atilika.kuromoji.Tokenizer').builder.build
+      else
+        @tokenizer = Rjb.import('org.atilika.kuromoji.Tokenizer').builder.userDictionary(user_dictionary).build
+      end
+    end
+
+    def tokenize(sentence)
+      process(:all_features, sentence)
+    end
+
+    def reading(sentence)
+      process(:getReading, sentence)
+    end
+
+    def tokenize_with_hash(sentence)
+      list = @tokenizer.tokenize(sentence)
+      iterator = list.iterator
+      result = []
+      while iterator.has_next
+        item = iterator.next
+        result << item.to_kuromoji_hash
+      end
+      result
+    end
+
+    def process(method, sentence)
+      list = @tokenizer.tokenize(sentence)
+      iterator = list.iterator
+      tokenized = {}
+      while iterator.has_next
+        item = iterator.next
+        tokenized[item.surface_form] = item.send(method)
+      end
+      tokenized
+    end
+  end
+end
